Refactor Gossip to Ducks
