--- client/actions/gossip_actions.js
@@ -1,27 +1,0 @@
-'use strict';
-
-const ADD_GOSSIP_START = 'ADD_GOSSIP_START';
-const ADD_GOSSIP_SUCCESS = 'ADD_GOSSIP_SUCCESS';
-const ADD_GOSSIP_FAIL = 'ADD_GOSSIP_FAIL';
-
-const DELETE_GOSSIP_START = 'DELETE_GOSSIP_START';
-const DELETE_GOSSIP_SUCCESS = 'DELETE_GOSSIP_SUCCESS';
-const DELETE_GOSSIP_FAIL = 'DELETE_GOSSIP_FAIL';
-
-const LOAD_MANY_GOSSIPS_START = 'LOAD_MANY_GOSSIPS_START';
-const LOAD_MANY_GOSSIPS_SUCCESS = 'LOAD_MANY_GOSSIPS_SUCCESS';
-const LOAD_MANY_GOSSIPS_FAIL = 'LOAD_MANY_GOSSIPS_FAIL';
-
-module.exports = {
-  ADD_GOSSIP_START,
-  ADD_GOSSIP_FAIL,
-  ADD_GOSSIP_SUCCESS,
-
-  DELETE_GOSSIP_START,
-  DELETE_GOSSIP_FAIL,
-  DELETE_GOSSIP_SUCCESS,
-
-  LOAD_MANY_GOSSIPS_FAIL,
-  LOAD_MANY_GOSSIPS_START,
-  LOAD_MANY_GOSSIPS_SUCCESS,
-};
