+++ client/reducers/reducers.js
@@ -6,7 +6,7 @@ const rootReducer = combineReducers({
   gacha: require('../modules/gacha').default,
   user: require('./user_reducer'),
   alert: require('../modules/alert').default,
-  gossip: require('./gossip_reducer'),
+  gossip: require('../modules/gossip').default,
   clock: require('./clock_reducer').clockReducer,
   dice: require('../modules/dice').default,
 });
